Refactor import paths for improved module organization
- Updated import statements across multiple modules to reflect the new structure under `utils/message` and `utils/system/core`, ensuring consistent access to message handling and global variables.
- Adjusted related UI and robot control modules to utilize the updated import paths, enhancing code clarity and maintainability.
- Enhanced the Python import helper with new paths for improved code completion related to message types and global variables.
